update ohkami v0.16 -> v0.17 & refactor some
