+++ frameworks/Rust/ohkami/src/main.rs
@@ -1,11 +1,14 @@
+mod fangs;
+use fangs::{SetServer, UsePostgres};
+
 mod models;
-pub use models::{Fortune, Message, World, WorldsQuery};
+use models::{Fortune, Message, World, WorldsQuery};
 
 mod postgres;
-pub use postgres::Postgres;
+use postgres::Postgres;
 
 mod templates;
-pub use templates::FortunesTemplate;
+use templates::FortunesTemplate;
 
 use ohkami::prelude::*;
 use ohkami::Memory;
@@ -13,16 +16,7 @@ use ohkami::Memory;
 
 #[tokio::main]
 async fn main() {
-    #[derive(Clone)]
-    struct SetServer;
-    impl FangAction for SetServer {
-        #[inline(always)]
-        async fn back<'a>(&'a self, res: &'a mut ohkami::Response) {
-            res.headers.set().Server("ohkami");
-        }
-    }
-
-    Ohkami::with((SetServer, Postgres::init().await), (
+    Ohkami::with((SetServer, UsePostgres::init().await), (
         "/json"     .GET(json_serialization),
         "/db"       .GET(single_database_query),
         "/queries"  .GET(multiple_database_query),
