+++ frameworks/Rust/ohkami/src/templates.rs
@@ -4,18 +4,20 @@ use crate::models::Fortune;
 
 
 #[derive(Template)]
-#[template(path="fortunes")]
+#[template(src = r#"<!DOCTYPE html><html><head><title>Fortunes</title></head><body><table><tr><th>id</th><th>message</th></tr>
+    {{~# each fortunes ~}}
+    <tr><td>{{id}}</td><td>{{message}}</td></tr>
+    {{~/each ~}}
+</table></body></html>"#)]
 pub struct FortunesTemplate {
     pub fortunes: Vec<Fortune>,
 }
+
 impl IntoResponse for FortunesTemplate {
     fn into_response(self) -> Response {
         match self.call() {
-            Ok(template) => Response::OK().html(template),
-            Err(error)   => {
-                eprintln!("Failed to render template: {error}");
-                Response::InternalServerError()
-            }
+            Ok(template) => Response::OK().with_html(template),
+            Err(_)       => Response::InternalServerError(),
         }
     }
 }
