--- frameworks/Rust/ohkami/templates/fortunes.hbs
@@ -1,5 +1,0 @@
-<!DOCTYPE html><html><head><title>Fortunes</title></head><body><table><tr><th>id</th><th>message</th></tr>
-      {{~# each fortunes ~}}
-      <tr><td>{{id}}</td><td>{{message}}</td></tr>
-      {{~/each ~}}
-</table></body></html>
